Use custom facts to get list of IP addresses.
There are several problem with the core-facts for ip-adresses. For
example, there is no way to get all IP for a interface that has more
than one IPv4 or more than one IPv6 address. And when I tried the new
better facter-3 (It's actually much better since it has structure data
not just strings), I got the link-local IPv6 address instead of the
global address.

 $ facter networking.ip6
 fe80::2e0:81ff:fed1:f409

This patch creates local fatcs to use as default by using
functions. From ruby class socket (ruby-2.1 is required), that can
list all IP for a interface, and can test if it's linklocal,
multicast or loopback.