Don't add facter values if lists are empty
The code in init.pp did replace empty list with standard fact
values. This was used as fallback if our custom fact failed for some
reason. This is no longer needed since the custom fact add the
standard fact itself (commit a14f1ab)

This had the unfortunately effect that in some circumstance it wasn't
possible to provide empty lists as parameters (issue #8)